more refactoring. cleaned up some of the db code, have some functional tests in test suite. removed extraneous code
