+++ src/server/db/data.js
@@ -1,0 +1,223 @@
+const dummyNames = [
+  "Liam",
+  "Emma",
+  "Noah",
+  "Olivia",
+  "William",
+  "Ava",
+  "James",
+  "Isabella",
+  "Oliver",
+  "Sophia",
+  "Benjamin",
+  "Charlotte",
+  "Elijah",
+  "Mia",
+  "Lucas",
+  "Amelia",
+  "Mason",
+  "Harper",
+  "Logan",
+  "Evelyn",
+  "Alexander",
+  "Abigail",
+  "Ethan",
+  "Emily",
+  "Jacob",
+  "Elizabeth",
+  "Michael",
+  "Mila",
+  "Daniel",
+  "Ella",
+  "Henry",
+  "Avery",
+  "Jackson",
+  "Sofia",
+  "Sebastian",
+  "Camila",
+  "Aiden",
+  "Aria",
+  "Matthew",
+  "Scarlett",
+  "Samuel",
+  "Victoria",
+  "David",
+  "Madison",
+  "Joseph",
+  "Luna",
+  "Carter",
+  "Grace",
+  "Owen",
+  "Chloe",
+  "Wyatt",
+  "Penelope",
+  "John",
+  "Layla",
+  "Jack",
+  "Riley",
+  "Luke",
+  "Zoey",
+  "Jayden",
+  "Nora",
+  "Dylan",
+  "Lily",
+  "Grayson",
+  "Elanor",
+  "Levi",
+  "Hannah",
+  "Issac",
+  "Lillian",
+  "Gabriel",
+  "Addison",
+  "Julian",
+  "Aubrey",
+  "Mateo",
+  "Ellie",
+  "Anthony",
+  "Stella",
+  "Jaxon",
+  "Natalie",
+  "Lincoln",
+  "Zoe",
+  "Joshua",
+  "Leah",
+  "Christopher",
+  "Hazel",
+  "Andrew",
+  "Violet",
+  "Theodore",
+  "Aurora",
+  "Caleb",
+  "Savannah",
+  "Ryan",
+  "Audrey",
+  "Asher",
+  "Brooklyn",
+  "Nathan",
+  "Bella",
+  "Thomas",
+  "Claire",
+  "Leo",
+  "Skylas",
+  "Isaiah",
+  "Lucy",
+  "Charles",
+  "Paisley",
+  "Josiah",
+  "Everly",
+  "Hudson",
+  "Annah",
+  "Christian",
+  "Caroline",
+  "Hunter",
+  "Nova",
+  "Connor",
+  "Genesis",
+  "Eli",
+  "Emilia",
+  "Ezra",
+  "Kennedy",
+  "Aaron",
+  "Samantha",
+  "Landon",
+  "Maya",
+  "Adrian",
+  "Willow",
+  "Jonathan",
+  "Kinsley",
+  "Nolan",
+  "Naomi",
+  "Jeremiah",
+  "Aaliyah",
+  "Easton",
+  "Elena"
+];
+
+const cities = [
+  "Montgomery, AL",
+  "Juneau, AK",
+  "Phoenix, AZ",
+  "Little Rock, AR",
+  "Sacramento, CA",
+  "Denver, CO",
+  "Hartford, CT",
+  "Dover, DE",
+  "Tallahassee, FL",
+  "Atlanta, GA",
+  "Honolulu, HI",
+  "Boise, ID",
+  "Springfield, IL",
+  "Indianapolis, IN",
+  "Des Moines, IA",
+  "Topeka, KS",
+  "Frankfort, KY",
+  "Baton Rouge, LA",
+  "Augusta, ME",
+  "Annapolis, MD",
+  "Boston, MA",
+  "Lansing, MI",
+  "Saint Paul, MN",
+  "Jackson, MS",
+  "Jefferson City, MO",
+  "Helena, MT",
+  "Lincoln, NE",
+  "Carson City, NV",
+  "Concord, NH",
+  "Trenton, NJ",
+  "Santa Fe, NM",
+  "Albany, NY",
+  "Raleigh, NC",
+  "Bismarck, ND",
+  "Columbus, OH",
+  "Oklahoma City, OK",
+  "Salem, OR",
+  "Harrisburg, PA",
+  "Providence, RI",
+  "Columbia, SC",
+  "Pierre, SD",
+  "Nashville, TN",
+  "Austin, TX",
+  "Salt Lake City, UT",
+  "Montpelier, VT",
+  "Richmond, VA",
+  "Olympia, WA",
+  "Charleston, WV",
+  "Madison, WI",
+  "Cheyenne, WY"
+];
+
+const positiveRevs = [
+  "WOW!",
+  "Love it!",
+  "No Complaints",
+  "Love at first sight",
+  "<3",
+  "Bought one for my wife too",
+  "Game Changer",
+  "Very Happy with this purchase",
+  "Super Fun"
+];
+
+const negativeRevs = [
+  "Waste of money",
+  "its okay but...",
+  "could definitely be improved",
+  "let down",
+  "avoid this product",
+  "Immediately returned!",
+  "Thumbs Down!",
+  "Not the best, not the worst",
+  "Go with Cannondale instead",
+  "Hard Pass!"
+];
+
+const dummyText =
+  "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
+
+module.exports = {
+  dummyNames,
+  cities,
+  positiveRevs,
+  negativeRevs,
+  dummyText
+};
